requirements.txt
Django==3.1.3
virtualenv==20.0.17
django-phone-field==1.8.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-files==1.1.0
mysqlclient==2.0.1
Pillow==8.0.1
django-upload-validator==1.1.5
django-imagekit==4.0.2
